chapel: apply fixes from upstream rel. 2.5 and 2.6
Apply the following improvements/fixes from upstream releases 2.5 and
2.6:
- https://github.com/chapel-lang/chapel/pull/26937
- https://github.com/chapel-lang/chapel/pull/27535
- https://github.com/chapel-lang/chapel/pull/27546
- https://github.com/chapel-lang/chapel/pull/27564
- https://github.com/chapel-lang/chapel/pull/27626
- https://github.com/chapel-lang/chapel/pull/27653

Additionally, disable gasnet+smp build and test on linux/arm64 to work
around an as yet undiagnosed build failure on the platform.

Includes a hack to allow gasnet+smp checkChplInstall testing to succeed
on a system with less than 4 cores, to be removed once there's an
upstream fix.

Adds a requirement of arm64 arch on macOS, as we've encountered
portability issues with gasnet+smp on x86_64 macOS 14, which we don't
have the time to fix at the moment and aren't worth denying the feature
to other platforms over.